packages/agents/requirements.txt
# Web server framework
fastapi
uvicorn[standard]

# Core LangChain packages (ensure compatible versions)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# New, dedicated package for Groq integration
langchain-groq

# New, dedicated package for HuggingFace embeddings
langchain-huggingface

# For embedding generation (the underlying library)
sentence-transformers

# Python client for Supabase
supabase

# To read .env file
python-dotenv

# For data validation (used by FastAPI and LangChain)
pydantic

# Direct PostgreSQL driver for our custom retriever
psycopg2-binary

# ... (existing requirements)
langgraph

langchain-openai